Fixed non-standard use of comparison operators
I've run a linter to catch non-standard uses of comparison operators.
Changes that have been made:

* Comparisons to True, False and None have all been changed to use the `is`
operator
* Comparisons to string literals have all been changed to use the `==`
oeprator
* Comparison to the empty list has been replaced with simply if my_list:
(which saves instantiating an empty list each time).
* Range comparisons (e.g. `if a < b and a > c`) have been changed to the
more standard `if c < a < b`.

I also caught one error in the numhess.rst which just required the
addition of a blank line to complete a code block so I've included it
here.