Automatic Sync from SCM-svn to GitHub syncing the following commits: RosaBulo: Rosa Bulo (REB) SCMSUITE-- SO107
I added arguments to the PLAMS get_conformations function, to more fully utilize the RDKit possibilities.
I added some restart options to the RDKit generator (mostly about saving geometries along the way).
I added some prints to stdout, to make it more clear why some candidates are not accepted by a conformers object.
I fixed a bug in the conformer code that keeps track of the indices of the conformers (stating that a conformer is the i-th added).