SO-- bug fix for logging 28-self._stdout_handler.level and csv logger in the jobmanager
having the same formatter for two handler is not ideal for a csv logger.
For csv formatter implemented there is an initialization step where you write down the headers, this in my opinion is best to have implemented in the formatter itself. But if both handlers have the same formatter the initialization step is called only once. therefore the Logger(ABC) now has _file_formatter and _stdout_formatter attributes